+++ src/components/EntitiesToolbar.vue
@@ -3,7 +3,7 @@
     <q-resize-observer @resize="resizeHandler"/>
     <slot name="selects"></slot>
     <div v-if="item">
-      <q-btn v-if="!item.deleted && mode !== undefined" flat dense class="on-right pull-right text-center rounded-borders q-px-xs q-py-none" color="white" @click="$emit('change:mode', !mode)" style="min-width: 73px; max-width: 73px;">
+      <q-btn v-if="!item.deleted && mode !== undefined" flat dense class="on-right pull-right text-center rounded-borders q-px-xs q-py-none" color="white" @click="$emit('change-mode', !mode)" style="min-width: 73px; max-width: 73px;">
         <q-icon size="1.5rem" color="white" :name="mode ? 'playlist_play' : 'history'"/>
         <div style="font-size: .7rem; line-height: .7rem">{{mode ? 'Real-time' : 'History'}}</div>
         <q-tooltip>Mode (Real-time/History)</q-tooltip>
@@ -16,7 +16,7 @@
         toggle-text-color="white"
         class="q-ml-sm gt-xs" size="sm"
         :value="ratio"
-        @input="r => $emit('change:ratio', r)"
+        @input="r => $emit('change-ratio', r)"
         :options="[{label: 'logs', value: 100},{label: 'both', value: 50},{label: 'messages', value: 0}]"
       />
     </div>
@@ -31,23 +31,23 @@
       </template>
     </div>
     <div v-else-if="item && (!$q.platform.is.desktop || width < 900)">
-      <q-btn flat icon="mdi-dots-vertical" color="white" v-if="hasActiveActions">
+      <q-btn flat icon="mdi-dots-vertical" color="white" v-if="hasActiveActions" round>
         <q-menu>
-          <q-list>
+          <q-list dark class="bg-grey-7">
             <q-item v-close-popup v-if="!item.deleted && ratio !== undefined">
               <q-btn-toggle
-                dense
-                color="grey-8"
-                toggle-color="white"
-                toggle-text-color="grey-9"
+                dense dark
+                color="grey-4"
+                toggle-color="grey-6"
+                toggle-text-color="white"
                 size="sm" flat
                 :value="ratio"
-                @input="r => $emit('change:ratio', r)"
+                @input="r => $emit('change-ratio', r)"
                 :options="[{label: 'logs', value: 100},{label: 'both', value: 50},{label: 'messages', value: 0}]"
               />
             </q-item>
             <template v-for="(action, index) in actions">
-              <q-item v-close-popup v-if="action.condition" clickable @click="action.handler" :key="index">
+              <q-item v-close-popup v-if="action.condition" clickable @click="action.handler" :key="index" dense>
                 <q-item-section avatar>
                   <q-icon :name="action.icon" />
                 </q-item-section>