+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [1.5.15] - 2020-10-29
+[rebase] added actions
+[feature] map on mabile view
+[feature] map on interval messages
 ### [1.5.14] - 2020-10-12
 [fix] naming new meta entities
 ### [1.5.13] - 2020-10-12